refactor(map): extract reference-marker rendering from CivicMap
The labelled reference dots (search/geolocate results, picked location) were
the one overlay built inline in CivicMap with innerHTML. Move them to
referenceMarkersLayer.renderReferenceMarkers(map, markers), built from DOM
nodes with the label set via textContent (so a label can't inject markup)
and colours set via style properties.
